split the reportsBoxOffciePage class and it's inner classes into seperate files
